Allows calc_orbit to take an array of keplerian elements as input. For calculations with 2-D arrays, currently looping through the "epochs" dimension (number of dates) because 1) it's likely to be smaller and 2) it prevents a lot of issues where sometimes we need to treat a keplerian element as a vector vs. a scalar.
The exception is the calculation of the eccentric anomaly, which is still done with a single value of eccentricity. Further upgrade changing this calculation can speed up the code.

For now, assume that the "mass" keyword still only points to one single value. Would not be hard to extend this to an array but not sure if that's what would be used? Let me know!

At the end, if np.squeeze() is used to reduce extra dimensions.