Update for results.py to fix the colorbar
The variable `index` that was producing an error was previously defined on line 333, but the code only works for the accepted inputs for cbar_param (sma1, ecc1, inc1, aop1, pan1, tau1, or any of these with the '1' switched for another number). This edit checks for whether the input parameters are of the accepted format and raises an error when they are not.